rename : jdbc 사용 위해 디렉토리 정리 및 이름 수정
